backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.4.1=pyhd3eb1b0_0
ca-certificates=2022.9.24=h033912b_0
certifi=2022.9.24=pyhd8ed1ab_0
django=3.2.5=pyhd3eb1b0_0
djangorestframework=3.14.0=pyhd8ed1ab_0
krb5=1.19.2=hcd88c3b_0
libcxx=14.0.6=h9765a3e_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libpq=12.9=h1c9f633_3
mysql-connector-c=6.1.11=h0f02589_1007
mysqlclient=2.0.3=py39h23ab428_1
ncurses=6.3=hca72f7f_3
openssl=1.1.1q=hfd90126_1
pip=22.2.2=py39hecd8cb5_0
psycopg2=2.8.6=py39hbcfaee0_1
python=3.9.13=hdfd78df_2
pytz=2022.1=py39hecd8cb5_0
readline=8.1.2=hca72f7f_1
setuptools=63.4.1=py39hecd8cb5_0
sqlite=3.39.3=h707629a_0
sqlparse=0.4.1=py_0
tk=8.6.12=h5d9f67b_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022e=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
zlib=1.2.12=h4dc903c_3
